Move cnl::make_signed and cnl::make_unsigned into separate header files
